+++ apps/nextjs/src/app/quiz-designer/preview/[slug]/preview.tsx
@@ -10,6 +10,10 @@ import HeroContainer from "@/components/HeroContainer";
 import { Icon } from "@/components/Icon";
 import Layout from "@/components/Layout";
 
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+
 export default function QuizPreview(questions) {
   const sharedData = questions;
   if (!sharedData) return null;
@@ -46,7 +50,9 @@ export default function QuizPreview(questions) {
                 </ul>
                 <p className="mt-14 font-semibold">Correct answer:</p>
                 <p>
-                  {question.answers.map((answer) => answer.value).join(", ")}
+                  {question.answers
+                    .map((answer: { value: any }) => answer.value)
+                    .join(", ")}
                 </p>
               </div>
             );