+++ packages/core/src/utils/textify.ts
@@ -1,12 +1,12 @@
 import yaml from "yaml";
 
 // Simplifies the input to a string for embedding
-export function textify(input: string | string[] | object): string {
+export function textify(input: unknown): string {
   if (Array.isArray(input)) {
     return input.map((row) => textify(row)).join("\n");
-  } else if (typeof input === "object") {
+  } else if (typeof input === "object" && input !== null) {
     return yaml.stringify(input);
   } else {
-    return input;
+    return String(input);
   }
 }