+++ packages/ingest/src/generate-lesson-plans/user-prompt-parts/exitQuiz.promptPart.ts
@@ -6,7 +6,7 @@ export function exitQuizPromptPart(rawLesson: RawLesson) {
 
   const exitQuizQuestions = exitQuiz ? transformQuiz(exitQuiz) : [];
   return exitQuizQuestions.length
-    ? `The lesson should include the following exit quiz questions. Include them within the lesson plan's exit quiz:
+    ? `The lesson should include the following exit quiz questions. Include them UNEDITED within the lesson plan's exit quiz:
 
 ${JSON.stringify(exitQuizQuestions)}`
     : null;